Undo extraction of _impl crate for simpler diff
