+++ faux_macros/src/methods/morphed.rs
