+++ faux_macros/src/methods/receiver.rs
