+++ faux_macros/src/self_type.rs
