--- faux_macros_impl/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "faux_macros_impl"
-version = "0.1.12"
-authors = ["Andres <andresnrx@gmail.com>"]
-edition = "2021"
-license = "MIT"
-description = "Implementations for #[create], #[methods], when!"
-homepage = "https://github.com/nrxus/faux"
-repository = "https://github.com/nrxus/faux"
-keywords = ["mock", "mocking", "test", "testing", "faux"]
-rust-version = "1.63"
-
-[dependencies]
-syn = { version = "2", features = ["full", "extra-traits"] }
-quote = "1"
-proc-macro2 = "1"
-darling = "0.20"
-uuid = { version = "1", features = ["v4"] }
-
-[dev-dependencies]
-faux = { path = "../" }
